2007-02-10 Erik Hjortsberg  <erik.hjortsberg@iteam.se>
	* src/components/ogre/EmberOgre.cpp: added a little bit more informative exception message
	* src/components/ogre/environment/caelum/src/SkyDome.cpp, src/components/ogre/environment/CaelumEnvironment.cpp: added extra checks for fragment and vertex programs
	* src/components/ogre/terrain/TerrainEditor.cpp: use correct type
	* src/components/ogre/widgets/EntityCEGUITexture.cpp: write to log before trying to create texture
	* src/components/ogre/scripting/bindings/lua/helpers/LuaConnector.h: fixed class/struct inconsistency
	* src/components/ogre/scripting/bindings/lua/helpers/LuaConnector.cpp: use correct type
	* src/components/ogre/jesus/XMLJesusSerializer.cpp: return statement required
	* src/services/config/ConfigService.cpp: MSVC likes std::string better
	* Ember.nsi: added Nullsoft installer file for Win32
	* Ember.bat: updated batch file

+++ src/components/ogre/scripting/bindings/lua/helpers/LuaConnector.cpp
@@ -66,7 +66,7 @@ void ConnectorBase::connect(const std::string & luaMethod)
 template <typename T0, typename T1, typename T2, typename T3> void ConnectorBase::callLuaMethod(T0 t0, T1 t1, T2 t2, T3 t3)
 {
 	try {
-		size_t numberOfArguments = mLuaTypeNames.size();
+		int numberOfArguments = static_cast<int>(mLuaTypeNames.size());
 		lua_State* state = EmberOgre::LuaConnector::getState();
 		
 		lua_getglobal(state, mLuaMethod.c_str());
